Bug 1311634 - Adopt more dockerflow patterns (#86)
- move to using NPM for frontend dependencies
- use official Python docker image
- get rid of pipstrap
- use correct user and group IDs

* Submit coverage from outside the container

* Use a app group to more closely follow Dockerflow

* Add Dockerfile.dev file

This does everything the main Dockerfile does except COPY’ing
the source code to /app. Instead we’re using a volume to mount it
under /app. We’re also creating /opt/coverage for the coverage
data created when running the tests.

This simplifies the docker-compose set up somewhat since we don’t
require seperation between development and testing environment
anymore that much. The only difference now is that the testing has
DEBUG set to False to more closely match the prod environment.

This also extended the run script to run the given command in case
it didn’t match any of the shortcut names such as “web”, “worker” or
“scheduler” so calls to docker-compose run still work as expected.

The run script will now also wait for the database and redis services
to start before continuing. That is to work around a missing feature
in docker-compose in which startup order isn’t defined.

* Run coverage without pytest-cov to be able to write coverage file to
location we want.

This adds a bin/test script to better handle Codecov's env var passing
with Docker.

* Remove --upgrade flag from pip install call

* Stop using netcat to wait for port to open

* Only submit coverage data in CI environment

* Run collectstatic during image building.

NOTE: This should be only merged when disabling the deployment time collectstatic.

* Add default CMD and ENTRYPOINT to prod Dockerfile.

* Add a new Build configuration and make sure the RQ jobs are only registered if Redis is available

* Specify jquery and moment-timezone as specific dependencies since eonasdan-bootstrap-datetimepicker changed things in https://github.com/Eonasdan/bootstrap-datetimepicker/commit/86b24d33a9b8b0b9473f95d9c54cb4322f4a1fc3

* Remove Circle CI in favor of Travis CI

This streamlines the Travis CI setup a lot.
Also adds IRC notifications for #atmo-dev room on irc.mozilla.org.