Add a Celery task for running a Spark job.
This task is used of Redbeat to schedule the Spark jobs using the Celery beat.
We add/remove Spark jobs from the schedule on save/delete and can restore the schedule from the database again.
This will also be useful for adding a “run now” button to Spark jobs.

This does a bunch of other things:

- send expiration emails for Spark jobs
- split tests into clusters and jobs packages with topic submodules (test_models, test_provisioners, test_views etc)
- refactor some test fixtures (e.g. remove all the pytest.mark.django_db decorators in favor of an autoused fixture)
- extend Celery task tests dramatically
- extend the test factories to cater to new tests
- fixed some model manager methods and test them
- refactor django.contrib.messages testing
- extend tests in various corner cases